Sub meta models

+++ GoalSetTracker/goal/views.py
@@ -16,12 +16,14 @@ def new_goal(request):
             finish_date = request.POST.get("finish_date")
             if finish_date == '':
                 finish_date = timezone.now()
-            user.goal_set.create(goal_text=request.POST.get("goal_text"),
+            goal = Goal.objects.create(goal_text=request.POST.get("goal_text"),
                                  finish_date=finish_date,
                                  create_date=timezone.now(),
                                  priority = request.POST.get("priority"),
-                                 state = request.POST.get("state")
+                                 state = request.POST.get("state"),
+                                 owner=user
                                  )
+            goal.save()
             return redirect_home(request.user.username)
         else:
             return render(request, 'goal/new_goal.html')
@@ -35,12 +37,16 @@ def new_sub_goal(request, goal_id):
             finish_date = request.POST.get("finish_date")
             if finish_date == '':
                 finish_date = timezone.now()
-            goal.subgoal_set.create(goal_text=request.POST.get("goal_text"),
-                                     finish_date=finish_date,
-                                     create_date=timezone.now(),
-                                     priority = request.POST.get("priority"),
-                                     state = request.POST.get("state")
-                                     )
+            if request.user == goal.owner:
+                subgoal = SubGoal.objects.create(
+                                        goal_text=request.POST.get("goal_text"),
+                                        finish_date=finish_date,
+                                        create_date=timezone.now(),
+                                        priority=request.POST.get("priority"),
+                                        state=request.POST.get("state"),
+                                        owner=request.user,
+                                        goal_father=goal
+                                        )
             return redirect_goal(goal_id)
         else:
             return render(request, 'goal/new_sub_goal.html', {'goal': goal})
@@ -50,11 +56,12 @@ def new_sub_goal(request, goal_id):
 def delete_goal(request, goal_id):
     goal = get_object_or_404(Goal, pk=goal_id)
     if request.user.is_authenticated:
-        if request.method == "POST":
-            goal.delete()
-            return redirect_home(request.user.username)
-        else:
-            return render(request, 'goal/delete.html', {'goal': goal})
+        if request.user == goal.owner:
+            if request.method == "POST":
+                goal.delete()
+                return redirect_home(request.user.username)
+            else:
+                return render(request, 'goal/delete.html', {'goal': goal})
     else:
         return HttpResponseRedirect("/login")
 
@@ -99,4 +106,3 @@ def redirect_goal(id, subgoal_id=None):
         return HttpResponseRedirect("/goal/{}/subgoal/{}".format(id, subgoal_id))
     else:
         return HttpResponseRedirect("/goal/{}".format(id))
-
